requirements.txt
flask
requests
werkzeug
python-docx
fpdf
pdfplumber
markdown
pytesseract
pillow
langdetect